pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seisCAE"
version = "0.1.0"
description = "Clustering seismic events using convolutional autoencoders"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Ankit Suthar", email = "ankit.suthar@example.com"}
]
keywords = ["seismology", "machine-learning", "clustering", "autoencoder", "deep-learning", "earthquake"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "matplotlib>=3.4.0",
    "pandas>=1.3.0",
    "obspy>=1.3.0",
    "torch>=1.9.0",
    "scikit-learn>=1.0.0",
    "seaborn>=0.11.0",
    "pyyaml>=5.4.0",
    "tqdm>=4.62.0",
    "click>=8.0.0",
    "umap-learn>=0.5.0",
    "tables>=3.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "isort>=5.10.0",
    "mypy>=0.950",
    "sphinx>=4.5.0",
    "sphinx-rtd-theme>=1.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/iSuthar/seisCAE"
Documentation = "https://github.com/iSuthar/seisCAE/blob/main/README.md"
Repository = "https://github.com/iSuthar/seisCAE"
"Bug Tracker" = "https://github.com/iSuthar/seisCAE/issues"

[project.scripts]
seiscae = "seiscae.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["seiscae*"]

[tool.setuptools.package-data]
seiscae = ["config/*.yaml"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=seiscae --cov-report=html --cov-report=term-missing"
